// etrack/etrack-analyst-portal-ui/src/app/@store/models/emailSend.ts
export class emailSend{
    emailCorrespondenceId: number | undefined;
    ccEmailId: String[] | undefined;
    emailBody: string | undefined;
    emailType :string | undefined;
    existingContents: String[] | undefined;
    fromEmailId: string | undefined ;
    subject: string | undefined;
    template: string | undefined;
    toEmailId :String[] | undefined;
    projectId:number| undefined;
    topicId:number |undefined;
    emailPurpose: string | undefined;
}